Project compiles, Product endpoints work except add-to-storage. Changed all id data types to BIGINT and LONG At least name validation works in ProductService at saveProductToDB()

+++ src/main/java/tarassov/project/dto/UserRequest.java
@@ -11,7 +11,7 @@ public class UserRequest {
     private String password;
     private String name;
     private String email;
-    private int productLimit;
+    private Long productLimit;
     private String userType;
     private Date birthDate;
